+++ instagram_to_discord/sites/twitter/twitter.py
@@ -1,7 +1,7 @@
 from typing import List
 from debug import DEBUG
 import discord
-from ..string_util import sophisticate_string
+from ...string_util import sophisticate_string
 
 async def send_twitter_images_from_cache_for_specified_index(
     skip_one: bool, image_urls: List[str], nums: List[int], message
